Move tests.vxlan plugin from global conftest to test specific folder
The vxlan specific fixtures are added to the global conftest.py as a plugin
`tests.vxlan`. It does not make sense.

This change removed the `tests.vxlan` plugin from global conftest.
The vxlan specific fixtures are moved to the vxlan/conftest.py file.
